Moved GigyaJSService to facade, exported in public API
